requirements.txt
aiohttp
python-socketio
pyttsx3